Manual upload of files
In effort to sync work done so far on this project across github.com, my local CLI, and local text editor applications with git integrations. It's important that no matter where I choose to work on this project, I can push my changes to github to keep track of and keep a record. My last attempt of this project I failed to do so and changes were never pushed, the initial commit was of many hours of previously completed work, and the record was no real record at all. Even know my initial commit yet again reflects a few hours of work and as-is already is functional, from hence-forth it is my intention to track all future changes as well as properly branch my playful ideas.